Issue/128 temp (#144)
* feat : 핸드폰 번호 회언가입

* refactor : 핸드폰번호 회원가입 API 수정

* feat : 로그인 API 추가

* refactor : api 500 에러 제외한 나머지 오류 isSuccess true로 변경

* :recycle: 게시글, 운동인증 작성 api 수정 (#130)

해시태그 리스트가 null이어도 문제 없도록 수정

* Issue/131 ♻️ 게시글, 운동인증 응답 시 해시태그에 # 제거, 💚 Nginx 설정 수정 (#132)

* :recycle: 게시글, 운동인증 응답 시 해시태그에 # 제거

* :green_heart: Nginx 설정 수정

업로드 파일 크기제한 수정

* :bug: 운동인증 API 버그 수정 (#134)

사진이 모두 지워지는 버그 수정

* :green_heart: Nginx 설정 변경 (#136)

cors 설정 변경

* :bug: Swagger 403에러 해결 (#138)

* issue/137 🐛 Swagger에서 요청 시 403에러 발생 해결 (#139)

* :bug: Swagger 403에러 해결

* :bug: Swaggre 403에러 해결

* 🐛 Swagger에서 요청 시 403에러 발생 해결 (#140)

* feat : 핸드폰 번호 회언가입

* refactor : 핸드폰번호 회원가입 API 수정

* feat : 로그인 API 추가

* refactor : api 500 에러 제외한 나머지 오류 isSuccess true로 변경

* :bug: Swagger 403에러 해결

* :bug: Swaggre 403에러 해결

* :bug: Swagger 403에러 해결

* :sparkles: 내가 적은 게시글, 인증 글 조회 API 추가 (#141)

* 코드 오류 수정

+++ src/main/java/fithub/app/service/impl/RecordServiceImpl.java
@@ -28,6 +28,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.io.IOException;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
@@ -59,11 +60,13 @@ public class RecordServiceImpl implements RecordService {
     public Record create(RecordRequestDto.CreateRecordDto request, User user, Integer categoryId) throws IOException
     {
         String exerciseTag = request.getExerciseTag();
-        HashTag exerciseHash = hashTagRepository.findByName('#' + exerciseTag).orElseGet(() -> HashTagConverter.newHashTag(exerciseTag));
-        List<HashTag> hashTagList = request.getHashTagList().stream()
-                .map(tag -> hashTagRepository.findByName('#' + tag).orElseGet(()-> HashTagConverter.newHashTag(tag)))
-                .collect(Collectors.toList());
-
+        HashTag exerciseHash = hashTagRepository.findByName(exerciseTag).orElseGet(() -> HashTagConverter.newHashTag(exerciseTag));
+        List<HashTag> hashTagList = new ArrayList<>();
+        if (request.getHashTagList() != null) {
+            hashTagList =request.getHashTagList().stream()
+                    .map(tag -> hashTagRepository.findByName(tag).orElseGet(() -> HashTagConverter.newHashTag(tag)))
+                    .collect(Collectors.toList());
+        }
         hashTagList.add(exerciseHash);
         Record record = RecordConverter.toRecord(request, user, hashTagList, categoryId);
         return recordRepository.save(record);
@@ -137,10 +140,10 @@ public class RecordServiceImpl implements RecordService {
         }
 
         String exerciseTag =  request.getExerciseTag();
-        HashTag exercisehashTag = hashTagRepository.findByName('#' + exerciseTag).orElseGet(() -> HashTagConverter.newHashTag(exerciseTag));
+        HashTag exercisehashTag = hashTagRepository.findByName(exerciseTag).orElseGet(() -> HashTagConverter.newHashTag(exerciseTag));
 
         List<HashTag> hashTagList = request.getHashTagList().stream()
-                .map(tag -> hashTagRepository.findByName('#' + tag).orElseGet(()-> HashTagConverter.newHashTag(tag)))
+                .map(tag -> hashTagRepository.findByName(tag).orElseGet(()-> HashTagConverter.newHashTag(tag)))
                 .collect(Collectors.toList());
 
         hashTagList.add(exercisehashTag);
@@ -149,21 +152,23 @@ public class RecordServiceImpl implements RecordService {
     }
 
     @Override
+    @Transactional(readOnly = false)
     public void deleteRecordSingle(Long recordId, User user) {
         Record record = recordRepository.findById(recordId).orElseThrow(() -> new RecordException(Code.RECORD_NOT_FOUND));
 
         if(!record.getUser().getId().equals(user.getId()))
             throw new RecordException(Code.RECORD_FORBIDDEN);
         String imageUrl = record.getImageUrl();
-        String Keyname = ArticleConverter.toKeyName(imageUrl);
-        amazonS3Manager.deleteFile(Keyname.substring(1));
+        if (imageUrl != null) {
+            String Keyname = ArticleConverter.toKeyName(imageUrl);
+            amazonS3Manager.deleteFile(Keyname.substring(1));
+        }
 
         List<RecordHashTag> recordHashTagList = record.getRecordHashTagList();
-
         for(int i = 0; i < recordHashTagList.size(); i++) {
-            RecordHashTag articleHashTag = recordHashTagList.get(i);
-            record.getRecordHashTagList().remove(articleHashTag);
-            recordHashTagRepository.delete(articleHashTag);
+            RecordHashTag recordHashTag = recordHashTagList.get(i);
+            record.getRecordHashTagList().remove(recordHashTag);
+            recordHashTagRepository.delete(recordHashTag);
         }
 
         if(recordHashTagList.size() > 0) {
